Migrate master branch to GitHub Pages
GitHub expects contents of the "public" directory to be published into
the master branch.  This is the only way how publishing of the user
pages can be handled within a single repository.  People normally put it
into a separate repository; however, I think it is more convinient to
create 'source' branch and make it default, while use the master branch
for publishing.

Normally this is going to be maintained by GitHub Actions CI automation,
but this is the first commit of the master branch.